Modified: Return value of heuristic, Added: Solution to heuristic function

+++ src/src/BFS.java
@@ -23,7 +23,7 @@ public class BFS extends Algorithm {
     }
     
     @Override
-    public int heuristic(Node node) {
+    public double heuristic(Node node) {
         return 0;
     }
