+++ src/src/DFS.java
@@ -108,7 +108,7 @@ public class DFS extends Algorithm
     }
 
     @Override
-    public int heuristic(Node node) {
+    public double heuristic(Node node) {
         return 0;
     }
